删除操作，传入 fileId 如果对应的是文件夹，是否删除文件夹？

+++ src/main/java/team/project/module/filestorage/internal/dao/LocalFileStorageDAO.java
@@ -42,7 +42,12 @@ public class LocalFileStorageDAO {
      * */
     public boolean delete(String filePath) {
         File file = new File(rootFolder,  filePath);
-        return file.delete(); /* <- 只要真的删除成功，才返回是 true，其他情况都是 false */
+        if ( ! file.exists() || file.isDirectory()) { /* <- 不删除文件夹，ljh_TODO 删除文件后，如果文件夹为空，是否需要删除文件夹？ */
+            return false;
+        }
+        else {
+            return file.delete(); /* <- 只要真的删除成功，才返回是 true，其他情况都是 false */
+        }
     }
 
     /* --------- */
